requirements.txt
alembic==1.3.1
amqp==2.5.2
Babel==2.7.0
bcrypt==3.1.7
billiard==3.6.1.0
blinker==1.4
celery==4.3.0
cffi==1.13.2
Click==7.0
colorama==0.4.3
decorator==4.4.1
Flask==1.1.1
Flask-BabelEx==0.9.3
Flask-Bcrypt==0.7.1
Flask-Components==0.1.1
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==20.0.4
infinity==1.4
intervals==0.8.1
itsdangerous==1.1.0
Jinja2==2.10.3
kombu==4.6.7
Mako==1.1.0
MarkupSafe==1.1.1
passlib==1.7.2
pycparser==2.19
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.3
redis==3.3.11
six==1.13.0
speaklater==1.3
SQLAlchemy==1.3.11
SQLAlchemy-Utils==0.36.0
validators==0.14.1
vine==1.3.0
Werkzeug==0.16.0
WTForms==2.2.1
WTForms-Alchemy==0.16.9
WTForms-Components==0.10.4
